requirements.txt
Flask==2.3.3
tensorflow==2.13.0
tensorflow-hub==0.14.0
numpy==1.24.3
soundfile==0.12.1
librosa==0.10.1
vosk==0.3.45
google-generativeai==0.3.2
twilio==8.10.0
requests==2.31.0
